+++ Library/SpdmLib/SpdmResponderLib/SpdmResponderLibHeartbeat.c
@@ -13,7 +13,7 @@ RETURN_STATUS
 EFIAPI
 SpdmGetResponseHeartbeat (
   IN     VOID                 *Context,
-  IN     UINT8                SessionId,
+  IN     UINT32               SessionId,
   IN     UINTN                RequestSize,
   IN     VOID                 *Request,
   IN OUT UINTN                *ResponseSize,
