+++ Library/SpdmLib/SpdmResponderLib/SpdmResponderLibPskFinish.c
@@ -55,7 +55,7 @@ RETURN_STATUS
 EFIAPI
 SpdmGetResponsePskFinish (
   IN     VOID                 *Context,
-  IN     UINT8                SessionId,
+  IN     UINT32               SessionId,
   IN     UINTN                RequestSize,
   IN     VOID                 *Request,
   IN OUT UINTN                *ResponseSize,
